+++ konfera/models/__init__.py
@@ -5,7 +5,7 @@ from .event import Event
 from .sponsor import Sponsor
 from .ticket_type import TicketType
 from .discount_codes import DiscountCodes
-from .attendee_ticket import Ticket
+from .ticket import Ticket
 from .speaker import Speaker
 from .talk import Talk
 from .room import Room
